backend/requirements.txt
audioread==3.0.1
blinker==1.9.0
celery==5.4.0
click==8.3.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg[binary]==3.2.12
python-dotenv==1.2.1
redis==5.2.0
soundfile==0.12.1
Werkzeug==3.1.3

